Compromise: Adds mono support to builds, but does not use MenuIterator on Windows
If this isn't good enough, then we might want to look over it again. If the menuiterator code isn't present, Mono doesn't display the menu and is unable to do anything. Because Mono and .NET builds are both bytecode-based, the binaries are shared, so ideally Windows .NET builds should work on Mono and visa versa - I tested this for the original menuiterator, but since it's not desired, this new compromise compiles it always and uses it for the Mono runtime only.